requirements.txt
torch
torchaudio
pyannote.audio
accelerate
kenlm
pysrt
pydub
python-dotenv
evaluate
pyctcdecode
fsspec